Adding some testing levels and corrected the character height to allow it to pass under platforms at his height high

+++ core/src/com/color/game/elements/BaseElement.java
@@ -24,8 +24,8 @@ public abstract class BaseElement extends Actor {
     }
 
     public Rectangle getBounds() {
-        int width = this.physicComponent.getUserData().getWidth();
-        int height = this.physicComponent.getUserData().getHeight();
+        float width = this.physicComponent.getUserData().getWidth();
+        float height = this.physicComponent.getUserData().getHeight();
         float x = (this.physicComponent.getBody().getPosition().x) - this.physicComponent.getUserData().getWidth()/2;
         float y = (this.physicComponent.getBody().getPosition().y) - this.physicComponent.getUserData().getHeight()/2;
         return new Rectangle(x * WORLD_TO_SCREEN, y * WORLD_TO_SCREEN, width * WORLD_TO_SCREEN, height * WORLD_TO_SCREEN);
