+++ core/src/com/color/game/elements/staticelements/platforms/ColorPlatform.java
@@ -63,10 +63,10 @@ public class ColorPlatform extends BaseStaticElement implements BaseColorPlatfor
         Color c = color.getColor();
         shapeRenderer.setColor(c.r, c.g, c.b, this.activated ? 1.0f : 0.5f);
 
-        int width = this.physicComponent.getUserData().getWidth() * WORLD_TO_SCREEN;
-        int height = this.physicComponent.getUserData().getHeight() * WORLD_TO_SCREEN;
-        int x = (int) (this.physicComponent.getBody().getPosition().x * WORLD_TO_SCREEN - width/2);
-        int y = (int) (this.physicComponent.getBody().getPosition().y * WORLD_TO_SCREEN - height/2);
+        float width = this.physicComponent.getUserData().getWidth() * WORLD_TO_SCREEN;
+        float height = this.physicComponent.getUserData().getHeight() * WORLD_TO_SCREEN;
+        float x = this.physicComponent.getBody().getPosition().x * WORLD_TO_SCREEN - width/2;
+        float y = this.physicComponent.getBody().getPosition().y * WORLD_TO_SCREEN - height/2;
 
         shapeRenderer.rect(x, y, width, height);
         shapeRenderer.end();
