+++ core/src/com/color/game/elements/userData/DynamicElementUserData.java
@@ -7,7 +7,7 @@ public class DynamicElementUserData extends UserData{
 
     private int contactsNumber;
 
-    public DynamicElementUserData(BaseDynamicElement element, int width, int height, UserDataType userDataType) {
+    public DynamicElementUserData(BaseDynamicElement element, float width, float height, UserDataType userDataType) {
         super(element, width, height, userDataType);
         contactsNumber = 0;
     }
